requirements.txt
# Ultralytics: YOLOv8을 가장 쉽게 사용할 수 있는 공식 라이브러리
ultralytics==8.3.204

# OpenCV: 비디오/이미지 처리를 위한 필수 라이브러리
opencv-python==4.12.0.88

# ===================================================================
# Data Handling & Smoothing
# ===================================================================
# NumPy: 모든 과학 및 수치 계산의 기반
numpy==2.2.6

# PyKalman: 칼만 필터를 쉽게 구현하기 위한 라이브러리
pykalman==0.10.2

# ===================================================================
# GUI Development
# ===================================================================
# PySide6: Qt 프레임워크를 파이썬에서 사용하기 위한 라이브러리 (LGPL 라이선스)
PySide6==6.9.3

# ===================================================================
# Video/Audio Handling
# ===================================================================
# FFmpeg Python Wrapper: 파이썬 코드에서 FFmpeg를 제어하기 위한 라이브러리
# 실제 FFmpeg 프로그램은 각 OS에 맞게 별도로 설치해야 합니다.
ffmpeg-python==0.2.0

tqdm==4.67.1            # Progress bars
python-dotenv==1.1.1   # Environment variable management

# ===================================================================
# AI & LLM
# ===================================================================
# Google Generative AI: Google의 생성형 AI (Gemini) API를 사용하기 위한 라이브러리
google-generativeai==0.8.5

# ===================================================================
# Audio Analysis (Phase 1)
# ===================================================================
# Librosa: 오디오 분석 라이브러리 (RMS, Spectral 분석)
librosa==0.11.0

# Soundfile: 오디오 파일 읽기/쓰기
soundfile==0.13.1

# OpenAI Whisper: 음성-텍스트 변환 (Apple Silicon MPS 지원)
openai-whisper==20250625

ultralytics
opencv-python
numpy
